+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Optional custom `$expected` parameter for `testDataValidationForeignKey()` and `testDataValidationIsUnique()`.
 
 ### Changed
-- **BREAKING CHANGE:** All type-specific/rule-dedicated methods now assert only their own validation rule for the field.
+- **BREAKING CHANGE:** All type-specific/rule-dedicated methods now assert only their own validation rule for the field, ignoring others.
 
 ### Fixed
 - Ignore dataValidation in `testDataValidationIsUnique()` to correctly assert build rules.